refine

+++ db/hashtable.h
@@ -35,7 +35,7 @@ class HashTable {
 	  uint32_t hash;	  // Hash of key(); used for fast sharding and comparisons	  
 	  uint32_t refs;
 	  
-	  size_t key_length;
+	  uint32_t key_length;
 	  char key_data[1];   // Beginning of key
   
 	  Slice key() const {
